Move impl methods not specific to type checking to live with the definition.
