+++ sway-core/src/language/ty/declaration/declaration.rs
@@ -1,8 +1,14 @@
 use std::fmt;
 
+use sway_error::error::CompileError;
 use sway_types::{Ident, Span, Spanned};
 
-use crate::{declaration_engine::*, error::*, language::ty::*, type_system::*};
+use crate::{
+    declaration_engine::*,
+    error::*,
+    language::{ty::*, Visibility},
+    type_system::*,
+};
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum TyDeclaration {
@@ -217,5 +223,248 @@ impl CollectTypesMetadata for TyDeclaration {
         } else {
             err(warnings, errors)
         }
+    }
+}
+
+impl TyDeclaration {
+    /// Retrieves the declaration as an enum declaration.
+    ///
+    /// Returns an error if `self` is not a [TyEnumDeclaration].
+    pub(crate) fn expect_enum(&self, access_span: &Span) -> CompileResult<TyEnumDeclaration> {
+        match self {
+            TyDeclaration::EnumDeclaration(decl_id) => {
+                CompileResult::from(de_get_enum(decl_id.clone(), access_span))
+            }
+            decl => err(
+                vec![],
+                vec![CompileError::DeclIsNotAnEnum {
+                    actually: decl.friendly_name().to_string(),
+                    span: decl.span(),
+                }],
+            ),
+        }
+    }
+
+    /// Retrieves the declaration as a struct declaration.
+    ///
+    /// Returns an error if `self` is not a [TyStructDeclaration].
+    pub(crate) fn expect_struct(&self, access_span: &Span) -> CompileResult<TyStructDeclaration> {
+        let mut warnings = vec![];
+        let mut errors = vec![];
+        match self {
+            TyDeclaration::StructDeclaration(decl_id) => {
+                let decl = check!(
+                    CompileResult::from(de_get_struct(decl_id.clone(), access_span)),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                ok(decl, warnings, errors)
+            }
+            decl => {
+                errors.push(CompileError::DeclIsNotAStruct {
+                    actually: decl.friendly_name().to_string(),
+                    span: decl.span(),
+                });
+                err(warnings, errors)
+            }
+        }
+    }
+
+    /// Retrieves the declaration as a function declaration.
+    ///
+    /// Returns an error if `self` is not a [TyFunctionDeclaration].
+    pub(crate) fn expect_function(
+        &self,
+        access_span: &Span,
+    ) -> CompileResult<TyFunctionDeclaration> {
+        let mut warnings = vec![];
+        let mut errors = vec![];
+        match self {
+            TyDeclaration::FunctionDeclaration(decl_id) => {
+                let decl = check!(
+                    CompileResult::from(de_get_function(decl_id.clone(), access_span)),
+                    return err(warnings, errors),
+                    warnings,
+                    errors,
+                );
+                ok(decl, warnings, errors)
+            }
+            decl => {
+                errors.push(CompileError::DeclIsNotAFunction {
+                    actually: decl.friendly_name().to_string(),
+                    span: decl.span(),
+                });
+                err(warnings, errors)
+            }
+        }
+    }
+
+    /// Retrieves the declaration as a variable declaration.
+    ///
+    /// Returns an error if `self` is not a [TyVariableDeclaration].
+    pub(crate) fn expect_variable(&self) -> CompileResult<&TyVariableDeclaration> {
+        let warnings = vec![];
+        let mut errors = vec![];
+        match self {
+            TyDeclaration::VariableDeclaration(decl) => ok(decl, warnings, errors),
+            decl => {
+                errors.push(CompileError::DeclIsNotAVariable {
+                    actually: decl.friendly_name().to_string(),
+                    span: decl.span(),
+                });
+                err(warnings, errors)
+            }
+        }
+    }
+
+    /// Retrieves the declaration as an Abi declaration.
+    ///
+    /// Returns an error if `self` is not a [TyAbiDeclaration].
+    pub(crate) fn expect_abi(&self, access_span: &Span) -> CompileResult<TyAbiDeclaration> {
+        match self {
+            TyDeclaration::AbiDeclaration(decl_id) => {
+                CompileResult::from(de_get_abi(decl_id.clone(), access_span))
+            }
+            decl => err(
+                vec![],
+                vec![CompileError::DeclIsNotAnAbi {
+                    actually: decl.friendly_name().to_string(),
+                    span: decl.span(),
+                }],
+            ),
+        }
+    }
+
+    /// friendly name string used for error reporting.
+    pub fn friendly_name(&self) -> &'static str {
+        use TyDeclaration::*;
+        match self {
+            VariableDeclaration(_) => "variable",
+            ConstantDeclaration(_) => "constant",
+            FunctionDeclaration(_) => "function",
+            TraitDeclaration(_) => "trait",
+            StructDeclaration(_) => "struct",
+            EnumDeclaration(_) => "enum",
+            ImplTrait { .. } => "impl trait",
+            AbiDeclaration(..) => "abi",
+            GenericTypeForFunctionScope { .. } => "generic type parameter",
+            ErrorRecovery => "error",
+            StorageDeclaration(_) => "contract storage declaration",
+        }
+    }
+
+    pub(crate) fn return_type(&self, access_span: &Span) -> CompileResult<TypeId> {
+        let mut warnings = vec![];
+        let mut errors = vec![];
+        let type_id = match self {
+            TyDeclaration::VariableDeclaration(decl) => decl.body.return_type,
+            TyDeclaration::FunctionDeclaration { .. } => {
+                errors.push(CompileError::Unimplemented(
+                    "Function pointers have not yet been implemented.",
+                    self.span(),
+                ));
+                return err(warnings, errors);
+            }
+            TyDeclaration::StructDeclaration(decl_id) => {
+                let decl = check!(
+                    CompileResult::from(de_get_struct(decl_id.clone(), &self.span())),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                decl.create_type_id()
+            }
+            TyDeclaration::EnumDeclaration(decl_id) => {
+                let decl = check!(
+                    CompileResult::from(de_get_enum(decl_id.clone(), access_span)),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                decl.create_type_id()
+            }
+            TyDeclaration::StorageDeclaration(decl_id) => {
+                let storage_decl = check!(
+                    CompileResult::from(de_get_storage(decl_id.clone(), &self.span())),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                insert_type(TypeInfo::Storage {
+                    fields: storage_decl.fields_as_typed_struct_fields(),
+                })
+            }
+            TyDeclaration::GenericTypeForFunctionScope { type_id, .. } => *type_id,
+            decl => {
+                errors.push(CompileError::NotAType {
+                    span: decl.span(),
+                    name: decl.to_string(),
+                    actually_is: decl.friendly_name(),
+                });
+                return err(warnings, errors);
+            }
+        };
+        ok(type_id, warnings, errors)
+    }
+
+    pub(crate) fn visibility(&self) -> CompileResult<Visibility> {
+        use TyDeclaration::*;
+        let mut warnings = vec![];
+        let mut errors = vec![];
+        let visibility = match self {
+            TraitDeclaration(decl_id) => {
+                let TyTraitDeclaration { visibility, .. } = check!(
+                    CompileResult::from(de_get_trait(decl_id.clone(), &decl_id.span())),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                visibility
+            }
+            ConstantDeclaration(decl_id) => {
+                let TyConstantDeclaration { visibility, .. } = check!(
+                    CompileResult::from(de_get_constant(decl_id.clone(), &decl_id.span())),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                visibility
+            }
+            StructDeclaration(decl_id) => {
+                let TyStructDeclaration { visibility, .. } = check!(
+                    CompileResult::from(de_get_struct(decl_id.clone(), &decl_id.span())),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                visibility
+            }
+            EnumDeclaration(decl_id) => {
+                let TyEnumDeclaration { visibility, .. } = check!(
+                    CompileResult::from(de_get_enum(decl_id.clone(), &decl_id.span())),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                visibility
+            }
+            FunctionDeclaration(decl_id) => {
+                let TyFunctionDeclaration { visibility, .. } = check!(
+                    CompileResult::from(de_get_function(decl_id.clone(), &decl_id.span())),
+                    return err(warnings, errors),
+                    warnings,
+                    errors
+                );
+                visibility
+            }
+            GenericTypeForFunctionScope { .. }
+            | ImplTrait { .. }
+            | StorageDeclaration { .. }
+            | AbiDeclaration(..)
+            | ErrorRecovery => Visibility::Public,
+            VariableDeclaration(decl) => decl.mutability.visibility(),
+        };
+        ok(visibility, warnings, errors)
     }
 }
