+++ sway-core/src/language/ty/declaration/struct.rs
@@ -1,8 +1,9 @@
 use std::hash::{Hash, Hasher};
 
+use sway_error::error::CompileError;
 use sway_types::{Ident, Span, Spanned};
 
-use crate::{language::Visibility, transform, type_system::*};
+use crate::{error::*, language::Visibility, transform, type_system::*};
 
 #[derive(Clone, Debug, Eq)]
 pub struct TyStructDeclaration {
@@ -60,6 +61,33 @@ impl MonomorphizeHelper for TyStructDeclaration {
 
     fn name(&self) -> &Ident {
         &self.name
+    }
+}
+
+impl TyStructDeclaration {
+    pub(crate) fn expect_field(&self, field_to_access: &Ident) -> CompileResult<&TyStructField> {
+        let warnings = vec![];
+        let mut errors = vec![];
+        match self
+            .fields
+            .iter()
+            .find(|TyStructField { name, .. }| name.as_str() == field_to_access.as_str())
+        {
+            Some(field) => ok(field, warnings, errors),
+            None => {
+                errors.push(CompileError::FieldNotFound {
+                    available_fields: self
+                        .fields
+                        .iter()
+                        .map(|TyStructField { name, .. }| name.to_string())
+                        .collect::<Vec<_>>()
+                        .join("\n"),
+                    field_name: field_to_access.clone(),
+                    struct_name: self.name.clone(),
+                });
+                err(warnings, errors)
+            }
+        }
     }
 }
