+++ sway-core/src/language/ty/expression/expression_variant.rs
@@ -591,4 +591,114 @@ impl TyExpressionVariant {
             _ => None,
         }
     }
+
+    /// recurse into `self` and get any return statements -- used to validate that all returns
+    /// do indeed return the correct type
+    /// This does _not_ extract implicit return statements as those are not control flow! This is
+    /// _only_ for explicit returns.
+    pub(crate) fn gather_return_statements(&self) -> Vec<&TyExpression> {
+        match self {
+            TyExpressionVariant::IfExp {
+                condition,
+                then,
+                r#else,
+            } => {
+                let mut buf = condition.gather_return_statements();
+                buf.append(&mut then.gather_return_statements());
+                if let Some(ref r#else) = r#else {
+                    buf.append(&mut r#else.gather_return_statements());
+                }
+                buf
+            }
+            TyExpressionVariant::CodeBlock(TyCodeBlock { contents, .. }) => {
+                let mut buf = vec![];
+                for node in contents {
+                    buf.append(&mut node.gather_return_statements())
+                }
+                buf
+            }
+            TyExpressionVariant::WhileLoop { condition, body } => {
+                let mut buf = condition.gather_return_statements();
+                for node in &body.contents {
+                    buf.append(&mut node.gather_return_statements())
+                }
+                buf
+            }
+            TyExpressionVariant::Reassignment(reassignment) => {
+                reassignment.rhs.gather_return_statements()
+            }
+            TyExpressionVariant::StorageReassignment(storage_reassignment) => {
+                storage_reassignment.rhs.gather_return_statements()
+            }
+            TyExpressionVariant::LazyOperator { lhs, rhs, .. } => [lhs, rhs]
+                .into_iter()
+                .flat_map(|expr| expr.gather_return_statements())
+                .collect(),
+            TyExpressionVariant::Tuple { fields } => fields
+                .iter()
+                .flat_map(|expr| expr.gather_return_statements())
+                .collect(),
+            TyExpressionVariant::Array { contents } => contents
+                .iter()
+                .flat_map(|expr| expr.gather_return_statements())
+                .collect(),
+            TyExpressionVariant::ArrayIndex { prefix, index } => [prefix, index]
+                .into_iter()
+                .flat_map(|expr| expr.gather_return_statements())
+                .collect(),
+            TyExpressionVariant::StructFieldAccess { prefix, .. } => {
+                prefix.gather_return_statements()
+            }
+            TyExpressionVariant::TupleElemAccess { prefix, .. } => {
+                prefix.gather_return_statements()
+            }
+            TyExpressionVariant::EnumInstantiation { contents, .. } => contents
+                .iter()
+                .flat_map(|expr| expr.gather_return_statements())
+                .collect(),
+            TyExpressionVariant::AbiCast { address, .. } => address.gather_return_statements(),
+            TyExpressionVariant::IntrinsicFunction(intrinsic_function_kind) => {
+                intrinsic_function_kind
+                    .arguments
+                    .iter()
+                    .flat_map(|expr| expr.gather_return_statements())
+                    .collect()
+            }
+            TyExpressionVariant::StructExpression { fields, .. } => fields
+                .iter()
+                .flat_map(|field| field.value.gather_return_statements())
+                .collect(),
+            TyExpressionVariant::FunctionApplication {
+                contract_call_params,
+                arguments,
+                selector,
+                ..
+            } => contract_call_params
+                .values()
+                .chain(arguments.iter().map(|(_name, expr)| expr))
+                .chain(
+                    selector
+                        .iter()
+                        .map(|contract_call_params| &*contract_call_params.contract_address),
+                )
+                .flat_map(|expr| expr.gather_return_statements())
+                .collect(),
+            TyExpressionVariant::EnumTag { exp } => exp.gather_return_statements(),
+            TyExpressionVariant::UnsafeDowncast { exp, .. } => exp.gather_return_statements(),
+
+            TyExpressionVariant::Return(exp) => {
+                vec![exp]
+            }
+            // if it is impossible for an expression to contain a return _statement_ (not an
+            // implicit return!), put it in the pattern below.
+            TyExpressionVariant::Literal(_)
+            | TyExpressionVariant::FunctionParameter { .. }
+            | TyExpressionVariant::AsmExpression { .. }
+            | TyExpressionVariant::VariableExpression { .. }
+            | TyExpressionVariant::AbiName(_)
+            | TyExpressionVariant::StorageAccess { .. }
+            | TyExpressionVariant::Break
+            | TyExpressionVariant::Continue => vec![],
+        }
+    }
 }
