+++ sway-core/src/semantic_analysis.rs
@@ -4,7 +4,7 @@ mod module;
 pub mod namespace;
 mod node_dependencies;
 mod program;
-mod storage_only_types;
+pub(crate) mod storage_only_types;
 mod type_check_context;
 
 pub(crate) use ast_node::*;
