+++ sway-core/src/semantic_analysis/ast_node/declaration.rs
@@ -15,261 +15,11 @@ pub use r#trait::*;
 pub use storage::*;
 
 use crate::{
-    declaration_engine::declaration_engine::*,
-    error::*,
     language::{ty, *},
     semantic_analysis::*,
-    type_system::*,
 };
 
-use sway_error::error::CompileError;
-use sway_types::{Span, Spanned};
-
-impl ty::TyDeclaration {
-    /// Retrieves the declaration as an enum declaration.
-    ///
-    /// Returns an error if `self` is not a [ty::TyEnumDeclaration].
-    pub(crate) fn expect_enum(&self, access_span: &Span) -> CompileResult<ty::TyEnumDeclaration> {
-        match self {
-            ty::TyDeclaration::EnumDeclaration(decl_id) => {
-                CompileResult::from(de_get_enum(decl_id.clone(), access_span))
-            }
-            decl => err(
-                vec![],
-                vec![CompileError::DeclIsNotAnEnum {
-                    actually: decl.friendly_name().to_string(),
-                    span: decl.span(),
-                }],
-            ),
-        }
-    }
-
-    /// Retrieves the declaration as a struct declaration.
-    ///
-    /// Returns an error if `self` is not a [ty::TyStructDeclaration].
-    pub(crate) fn expect_struct(
-        &self,
-        access_span: &Span,
-    ) -> CompileResult<ty::TyStructDeclaration> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        match self {
-            ty::TyDeclaration::StructDeclaration(decl_id) => {
-                let decl = check!(
-                    CompileResult::from(de_get_struct(decl_id.clone(), access_span)),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                ok(decl, warnings, errors)
-            }
-            decl => {
-                errors.push(CompileError::DeclIsNotAStruct {
-                    actually: decl.friendly_name().to_string(),
-                    span: decl.span(),
-                });
-                err(warnings, errors)
-            }
-        }
-    }
-
-    /// Retrieves the declaration as a function declaration.
-    ///
-    /// Returns an error if `self` is not a [ty::TyFunctionDeclaration].
-    pub(crate) fn expect_function(
-        &self,
-        access_span: &Span,
-    ) -> CompileResult<ty::TyFunctionDeclaration> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        match self {
-            ty::TyDeclaration::FunctionDeclaration(decl_id) => {
-                let decl = check!(
-                    CompileResult::from(de_get_function(decl_id.clone(), access_span)),
-                    return err(warnings, errors),
-                    warnings,
-                    errors,
-                );
-                ok(decl, warnings, errors)
-            }
-            decl => {
-                errors.push(CompileError::DeclIsNotAFunction {
-                    actually: decl.friendly_name().to_string(),
-                    span: decl.span(),
-                });
-                err(warnings, errors)
-            }
-        }
-    }
-
-    /// Retrieves the declaration as a variable declaration.
-    ///
-    /// Returns an error if `self` is not a [ty::TyVariableDeclaration].
-    pub(crate) fn expect_variable(&self) -> CompileResult<&ty::TyVariableDeclaration> {
-        let warnings = vec![];
-        let mut errors = vec![];
-        match self {
-            ty::TyDeclaration::VariableDeclaration(decl) => ok(decl, warnings, errors),
-            decl => {
-                errors.push(CompileError::DeclIsNotAVariable {
-                    actually: decl.friendly_name().to_string(),
-                    span: decl.span(),
-                });
-                err(warnings, errors)
-            }
-        }
-    }
-
-    /// Retrieves the declaration as an Abi declaration.
-    ///
-    /// Returns an error if `self` is not a [ty::TyAbiDeclaration].
-    pub(crate) fn expect_abi(&self, access_span: &Span) -> CompileResult<ty::TyAbiDeclaration> {
-        match self {
-            ty::TyDeclaration::AbiDeclaration(decl_id) => {
-                CompileResult::from(de_get_abi(decl_id.clone(), access_span))
-            }
-            decl => err(
-                vec![],
-                vec![CompileError::DeclIsNotAnAbi {
-                    actually: decl.friendly_name().to_string(),
-                    span: decl.span(),
-                }],
-            ),
-        }
-    }
-
-    /// friendly name string used for error reporting.
-    pub fn friendly_name(&self) -> &'static str {
-        use ty::TyDeclaration::*;
-        match self {
-            VariableDeclaration(_) => "variable",
-            ConstantDeclaration(_) => "constant",
-            FunctionDeclaration(_) => "function",
-            TraitDeclaration(_) => "trait",
-            StructDeclaration(_) => "struct",
-            EnumDeclaration(_) => "enum",
-            ImplTrait { .. } => "impl trait",
-            AbiDeclaration(..) => "abi",
-            GenericTypeForFunctionScope { .. } => "generic type parameter",
-            ErrorRecovery => "error",
-            StorageDeclaration(_) => "contract storage declaration",
-        }
-    }
-
-    pub(crate) fn return_type(&self, access_span: &Span) -> CompileResult<TypeId> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let type_id = match self {
-            ty::TyDeclaration::VariableDeclaration(decl) => decl.body.return_type,
-            ty::TyDeclaration::FunctionDeclaration { .. } => {
-                errors.push(CompileError::Unimplemented(
-                    "Function pointers have not yet been implemented.",
-                    self.span(),
-                ));
-                return err(warnings, errors);
-            }
-            ty::TyDeclaration::StructDeclaration(decl_id) => {
-                let decl = check!(
-                    CompileResult::from(de_get_struct(decl_id.clone(), &self.span())),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                decl.create_type_id()
-            }
-            ty::TyDeclaration::EnumDeclaration(decl_id) => {
-                let decl = check!(
-                    CompileResult::from(de_get_enum(decl_id.clone(), access_span)),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                decl.create_type_id()
-            }
-            ty::TyDeclaration::StorageDeclaration(decl_id) => {
-                let storage_decl = check!(
-                    CompileResult::from(de_get_storage(decl_id.clone(), &self.span())),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                insert_type(TypeInfo::Storage {
-                    fields: storage_decl.fields_as_typed_struct_fields(),
-                })
-            }
-            ty::TyDeclaration::GenericTypeForFunctionScope { type_id, .. } => *type_id,
-            decl => {
-                errors.push(CompileError::NotAType {
-                    span: decl.span(),
-                    name: decl.to_string(),
-                    actually_is: decl.friendly_name(),
-                });
-                return err(warnings, errors);
-            }
-        };
-        ok(type_id, warnings, errors)
-    }
-
-    pub(crate) fn visibility(&self) -> CompileResult<Visibility> {
-        use ty::TyDeclaration::*;
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let visibility = match self {
-            TraitDeclaration(decl_id) => {
-                let ty::TyTraitDeclaration { visibility, .. } = check!(
-                    CompileResult::from(de_get_trait(decl_id.clone(), &decl_id.span())),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                visibility
-            }
-            ConstantDeclaration(decl_id) => {
-                let ty::TyConstantDeclaration { visibility, .. } = check!(
-                    CompileResult::from(de_get_constant(decl_id.clone(), &decl_id.span())),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                visibility
-            }
-            StructDeclaration(decl_id) => {
-                let ty::TyStructDeclaration { visibility, .. } = check!(
-                    CompileResult::from(de_get_struct(decl_id.clone(), &decl_id.span())),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                visibility
-            }
-            EnumDeclaration(decl_id) => {
-                let ty::TyEnumDeclaration { visibility, .. } = check!(
-                    CompileResult::from(de_get_enum(decl_id.clone(), &decl_id.span())),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                visibility
-            }
-            FunctionDeclaration(decl_id) => {
-                let ty::TyFunctionDeclaration { visibility, .. } = check!(
-                    CompileResult::from(de_get_function(decl_id.clone(), &decl_id.span())),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                visibility
-            }
-            GenericTypeForFunctionScope { .. }
-            | ImplTrait { .. }
-            | StorageDeclaration { .. }
-            | AbiDeclaration(..)
-            | ErrorRecovery => Visibility::Public,
-            VariableDeclaration(decl) => decl.mutability.visibility(),
-        };
-        ok(visibility, warnings, errors)
-    }
-}
+use sway_types::Spanned;
 
 impl ty::TyTraitFn {
     /// This function is used in trait declarations to insert "placeholder" functions
