+++ sway-core/src/semantic_analysis/ast_node/declaration/struct.rs
@@ -4,8 +4,6 @@ use crate::{
     semantic_analysis::*,
     type_system::*,
 };
-use sway_error::error::CompileError;
-use sway_types::Ident;
 
 impl ty::TyStructDeclaration {
     pub(crate) fn type_check(
@@ -63,34 +61,6 @@ impl ty::TyStructDeclaration {
         };
 
         ok(decl, warnings, errors)
-    }
-
-    pub(crate) fn expect_field(
-        &self,
-        field_to_access: &Ident,
-    ) -> CompileResult<&ty::TyStructField> {
-        let warnings = vec![];
-        let mut errors = vec![];
-        match self
-            .fields
-            .iter()
-            .find(|ty::TyStructField { name, .. }| name.as_str() == field_to_access.as_str())
-        {
-            Some(field) => ok(field, warnings, errors),
-            None => {
-                errors.push(CompileError::FieldNotFound {
-                    available_fields: self
-                        .fields
-                        .iter()
-                        .map(|ty::TyStructField { name, .. }| name.to_string())
-                        .collect::<Vec<_>>()
-                        .join("\n"),
-                    field_name: field_to_access.clone(),
-                    struct_name: self.name.clone(),
-                });
-                err(warnings, errors)
-            }
-        }
     }
 }
