+++ sway-core/src/semantic_analysis/ast_node/mod.rs
@@ -10,11 +10,7 @@ pub(crate) use mode::*;
 use crate::{
     declaration_engine::{declaration_engine::*, DeclarationId},
     error::*,
-    language::{
-        parsed::*,
-        ty::{self, TyExpression},
-        Visibility,
-    },
+    language::{parsed::*, ty, Visibility},
     semantic_analysis::*,
     type_system::*,
     types::DeterministicallyAborts,
@@ -28,85 +24,6 @@ use sway_error::{
 use sway_types::{span::Span, state::StateIndex, style::is_screaming_snake_case, Spanned};
 
 impl ty::TyAstNode {
-    /// Returns `true` if this AST node will be exported in a library, i.e. it is a public declaration.
-    pub(crate) fn is_public(&self) -> CompileResult<bool> {
-        use ty::TyAstNodeContent::*;
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let public = match &self.content {
-            Declaration(decl) => {
-                let visibility = check!(
-                    decl.visibility(),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                visibility.is_public()
-            }
-            Expression(_) | SideEffect | ImplicitReturnExpression(_) => false,
-        };
-        ok(public, warnings, errors)
-    }
-
-    /// Naive check to see if this node is a function declaration of a function called `main` if
-    /// the [TreeType] is Script or Predicate.
-    pub(crate) fn is_main_function(&self, tree_type: TreeType) -> CompileResult<bool> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        match &self {
-            ty::TyAstNode {
-                span,
-                content:
-                    ty::TyAstNodeContent::Declaration(ty::TyDeclaration::FunctionDeclaration(decl_id)),
-                ..
-            } => {
-                let ty::TyFunctionDeclaration { name, .. } = check!(
-                    CompileResult::from(de_get_function(decl_id.clone(), span)),
-                    return err(warnings, errors),
-                    warnings,
-                    errors
-                );
-                let is_main = name.as_str() == sway_types::constants::DEFAULT_ENTRY_POINT_FN_NAME
-                    && matches!(tree_type, TreeType::Script | TreeType::Predicate);
-                ok(is_main, warnings, errors)
-            }
-            _ => ok(false, warnings, errors),
-        }
-    }
-
-    /// recurse into `self` and get any return statements -- used to validate that all returns
-    /// do indeed return the correct type
-    /// This does _not_ extract implicit return statements as those are not control flow! This is
-    /// _only_ for explicit returns.
-    pub(crate) fn gather_return_statements(&self) -> Vec<&TyExpression> {
-        match &self.content {
-            ty::TyAstNodeContent::ImplicitReturnExpression(ref exp) => {
-                exp.gather_return_statements()
-            }
-            // assignments and  reassignments can happen during control flow and can abort
-            ty::TyAstNodeContent::Declaration(ty::TyDeclaration::VariableDeclaration(decl)) => {
-                decl.body.gather_return_statements()
-            }
-            ty::TyAstNodeContent::Expression(exp) => exp.gather_return_statements(),
-            ty::TyAstNodeContent::SideEffect | ty::TyAstNodeContent::Declaration(_) => vec![],
-        }
-    }
-
-    fn type_info(&self) -> TypeInfo {
-        // return statement should be ()
-        use ty::TyAstNodeContent::*;
-        match &self.content {
-            Declaration(_) => TypeInfo::Tuple(Vec::new()),
-            Expression(ty::TyExpression { return_type, .. }) => {
-                crate::type_system::look_up_type_id(*return_type)
-            }
-            ImplicitReturnExpression(ty::TyExpression { return_type, .. }) => {
-                crate::type_system::look_up_type_id(*return_type)
-            }
-            SideEffect => TypeInfo::Tuple(Vec::new()),
-        }
-    }
-
     pub(crate) fn type_check(mut ctx: TypeCheckContext, node: AstNode) -> CompileResult<Self> {
         let mut warnings = Vec::new();
         let mut errors = Vec::new();
